+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 2.9.5
 - Fix #152 Cookie on localhost/ (root path) and IE in Warehouse.php
 - Add jQuery Chosen 1.5.1 plugin in assets/js/jquery-chosen
 - Add ChosenSelectInput() function in Inputs.php
+- Add "Limit Existing Contacts & Addresses to current school" global setting in Configuration.php (sponsored by Aptiris)
 
 Changes in 2.9.4
 ----------------